# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NEPZIA Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /listings/{id}/offers:
    post:
      operationId: makeOffer
      tags: [offers]
      summary: Make an offer on a listing
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OfferInput"
      responses:
        "201":
          description: Offer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"

  # ─── My Listings ────────────────────────────────────────────────────────────
...
  /offers/{id}/respond:
    patch:
      operationId: respondToOffer
      tags: [offers]
      summary: Accept, reject, or counter an offer
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OfferResponse"
      responses:
        "200":
          description: Updated offer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"

  # ─── Categories ─────────────────────────────────────────────────────────────
...
  /wishlist/{listingId}:
    post:
      operationId: addToWishlist
      tags: [wishlist]
      summary: Add listing to wishlist
      parameters:
        - name: listingId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "201":
          description: Added
    delete:
      operationId: removeFromWishlist
      tags: [wishlist]
      summary: Remove listing from wishlist
      parameters:
        - name: listingId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Removed

  # ─── Messages ────────────────────────────────────────────────────────────────
  /conversations:
    get:
      operationId: listConversations
      tags: [messages]
      summary: List user's conversations
      responses:
        "200":
          description: Conversations list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
    post:
      operationId: createConversation
      tags: [messages]
      summary: Start a conversation about a listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConversationInput"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
  /conversations/{id}/messages:
    get:
      operationId: getMessages
      tags: [messages]
      summary: Get messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  # ─── Users / Sellers ────────────────────────────────────────────────────────
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a single notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Done

  # ─── Stats / Wow ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    OfferInput:
      type: object
      required: [amount]
      properties:
        amount: { type: number }
        message: { type: string }
    OfferResponse:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [accepted, rejected, countered]
        counterAmount: { type: number }
    Offer:
      type: object
      required: [id, listingId, buyerId, sellerId, amount, status, createdAt]
      properties:
        id: { type: integer }
        listingId: { type: integer }
        listingTitle: { type: ["string", "null"] }
        listingImage: { type: ["string", "null"] }
        buyerId: { type: string }
        buyerName: { type: ["string", "null"] }
        sellerId: { type: string }
        amount: { type: number }
        message: { type: ["string", "null"] }
        status: { type: string }
        counterAmount: { type: ["number", "null"] }
        createdAt: { type: string }
...
    Conversation:
      type: object
      required: [id, listingId, buyerId, sellerId, createdAt]
      properties:
        id: { type: integer }
        listingId: { type: integer }
        listingTitle: { type: ["string", "null"] }
        listingImage: { type: ["string", "null"] }
        buyerId: { type: string }
        sellerId: { type: string }
        otherPartyName: { type: ["string", "null"] }
        otherPartyAvatar: { type: ["string", "null"] }
        lastMessage: { type: ["string", "null"] }
        lastMessageAt: { type: ["string", "null"] }
        unreadCount: { type: integer }
        createdAt: { type: string }
    ConversationInput:
      type: object
      required: [listingId, sellerId]
      properties:
        listingId: { type: integer }
        sellerId: { type: string }
        initialMessage: { type: string }
    Message:
      type: object
      required: [id, conversationId, senderId, content, createdAt]
      properties:
        id: { type: integer }
        conversationId: { type: integer }
        senderId: { type: string }
        content: { type: string }
        createdAt: { type: string }
    MessageInput:
      type: object
      required: [content]
      properties:
        content: { type: string }
...