# lib/api-zod/src/generated/types/offerResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NEPZIA Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { OfferResponseAction } from './offerResponseAction';

export interface OfferResponse {
  action: OfferResponseAction;
  counterAmount?: number;
}