# lib/api-zod/src/src/generated/types/offer.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NEPZIA Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface Offer {
  id: number;
  listingId: number;
  /** @nullable */
  listingTitle?: string | null;
  /** @nullable */
  listingImage?: string | null;
  buyerId: string;
  /** @nullable */
  buyerName?: string | null;
  sellerId: string;
  amount: number;
  /** @nullable */
  message?: string | null;
  status: string;
  /** @nullable */
  counterAmount?: number | null;
  createdAt: string;
}